+++ src/expression/function/compile.js
@@ -4,9 +4,9 @@ import { deepMap } from '../../utils/collection'
 import { factory } from '../../utils/factory'
 
 const name = 'compile'
-const dependencies = ['typed', 'expression.parse']
+const dependencies = ['typed', 'parse']
 
-export const createCompile = /* #__PURE__ */ factory(name, dependencies, ({ typed, expression: { parse } }) => {
+export const createCompile = /* #__PURE__ */ factory(name, dependencies, ({ typed, parse }) => {
   /**
    * Parse and compile an expression.
    * Returns a an object with a function `evaluate([scope])` to evaluate the