+++ HISTORY.md
@@ -16,6 +16,9 @@ Breaking changes:
   recreates all affected functions when the configuration changes.
 - Deprecated the `Matrix.storage` function. Use `math.matrix` instead to create
   a matrix.
+- Deprecated function `math.expression.parse`, use `math.parse` instead.
+  Was used before for example to customize supported characters by replacing
+  `math.parse.isAlpha`.
 
 Non breaking: